requirements.txt
python-dotenv
browser-use